// src/types/validate-params.enum.ts
export enum FilmTitleValid {
  MinLength = 2,
  MaxLength = 100
}

export enum FilmDescriptionValid {
  MinLength = 20,
  MaxLength = 1024
}

export enum FilmDirectorValid {
  MinLength = 2,
  MaxLength = 50
}

export enum UserPasswordValid {
  MinLength = 6,
  MaxLength = 12
}

export enum CommentTextValid {
  MinLength = 5,
  MaxLength = 1024
}

export enum CommentRatingValid {
  MinValue = 1,
  MaxValue = 10
}